fix: refactor producer
